requirements.txt
fastapi
uvicorn
starlette
requests
Pillow